remove popup.js
