+++ ui/sidepanel/sidepanel.html
@@ -65,6 +65,6 @@
     <h2>Quiz</h2>
     <div id="quiz"></div>
     <button id="submit">Submit</button>
-    <script type="module" src="../popup/popup.js"></script>
+    <script type="module" src="./sidepanel.js"></script>
   </body>
 </html>
